requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.15.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.1
protobuf==5.29.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.9.0
streamlit==1.40.2
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
websockets==13.1
yarl==1.18.3
